refactor: tooltip into component
