+++ src/components/analytics/TaxaTreemap.tsx
@@ -1,6 +1,7 @@
 import { ResponsiveContainer, Tooltip, Treemap } from "recharts";
 
 import { TaxaTreemapCell } from "@/components/analytics/TaxaTreemapCell";
+import { TaxaTreemapTooltip } from "@/components/analytics/TaxaTreemapTooltip";
 import { useFilters } from "@/lib/hooks/useFilters";
 import { useTaxaCount } from "@/lib/hooks/useTaxaCount";
 
@@ -22,7 +23,6 @@ interface TreemapItem {
 }
 
 const PRIMARY_TAXA_LIMIT = 8;
-const OTHER_PREVIEW_LIMIT = 8;
 
 function getTreemapSize(count: number): number {
   return Math.log1p(count);
@@ -103,53 +103,6 @@ function TaxaTreemap({ deploymentId }: TaxaTreemapProps) {
   };
 
   return content();
-}
-
-function TaxaTreemapTooltip({
-  active,
-  payload,
-}: {
-  active?: boolean;
-  payload?: Array<{ payload: TreemapItem }>;
-}) {
-  if (!active || !payload?.[0]?.payload) {
-    return null;
-  }
-
-  const item = payload[0].payload;
-  const otherItems = item.items ?? [];
-
-  return (
-    <div
-      className="rounded-md border px-3 py-2 shadow-lg"
-      style={{
-        backgroundColor: "var(--color-card)",
-        borderColor: "var(--color-border)",
-        color: "var(--color-foreground)",
-      }}
-    >
-      <p className="text-sm font-semibold">{item.name}</p>
-      <p className="text-xs text-muted-foreground">{item.count} detections</p>
-      {item.name === "Others" && otherItems.length > 0 && (
-        <div className="mt-2 border-t border-border pt-2">
-          <p className="mb-1 text-xs font-medium text-muted-foreground">Includes</p>
-          <div className="flex flex-col gap-1 text-xs">
-            {otherItems.slice(0, OTHER_PREVIEW_LIMIT).map((otherItem) => (
-              <div key={otherItem.name} className="flex items-center justify-between gap-4">
-                <span className="truncate">{otherItem.name}</span>
-                <span className="shrink-0 text-muted-foreground">{otherItem.count}</span>
-              </div>
-            ))}
-            {otherItems.length > OTHER_PREVIEW_LIMIT && (
-              <p className="pt-1 text-xs text-muted-foreground">
-                +{otherItems.length - OTHER_PREVIEW_LIMIT} more taxa
-              </p>
-            )}
-          </div>
-        </div>
-      )}
-    </div>
-  );
 }
 
 export { TaxaTreemap };
